Neils dev

+++ springboot/src/main/java/com/qa/project/persistence/repo/BuyerRepo.java
@@ -1,9 +1,11 @@
 package com.qa.project.persistence.repo;
 
-import com.qa.project.persistence.domain.BuyerDomain;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import com.qa.project.persistence.domain.BuyerDomain;
+
 @Repository
-public interface BuyerRepo extends JpaRepository<BuyerDomain, Long> {
+public interface BuyerRepo extends JpaRepository<BuyerDomain, Integer> {
+
 }
