+++ springboot/src/main/java/com/qa/project/persistence/repo/SellerRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.qa.project.persistence.domain.SellerDomain;
 
 @Repository
-public interface SellerRepo extends JpaRepository<SellerDomain, Long> {
+public interface SellerRepo extends JpaRepository<SellerDomain, Integer> {
 
 }
